* master:
  Fix formatting for failed validation for one of the items in ListField (#413)
  PyTest: Run failed test and then new test cases first (#414)
  update dependencies to latest (#412)
  Add "django.contrib.postgres" to list of INSTALLED_APPS
  docs: improve setup docs on mac
  docs(): automate generating the latest db schema on deployment in docs (#408)
  update readme
  Add artwork
  chore - upgrade django (#407)
  feat: upgrade python dependencies to latest stable (#406)
  Upgrade django 3.0.5 -> 3.0.6
  Update documentation (#405)
  Remove web pack + replace fabric with makefile (#404)

+++ run_test.sh
@@ -5,12 +5,14 @@ err=0
 trap 'err=1' ERR
 
 # Clean
-if psql -lqt | cut -d \| -f 1 | grep -qw hello_world ; then
-    read -p "Database 'hello_world' required for running the tests already exist. Do you want to delete it (y)?" yn
-    if echo "$yn" | grep -iq "^n" ;then
-        exit
-    else
-        dropdb hello_world
+if [ ! $CI ]; then
+    if psql -lqt | cut -d \| -f 1 | grep -qw hello_world ; then
+        read -p "Database 'hello_world' required for running the tests already exist. Do you want to delete it (y)?" yn
+        if echo "$yn" | grep -iq "^n" ;then
+            exit
+        else
+            dropdb hello_world
+        fi
     fi
 fi
 
@@ -21,10 +23,9 @@ yes 'y' | cookiecutter . --no-input
 
 # Run the tests present inside generate project
 cd hello-world-web;
-npm run build
 source venv/bin/activate
 ansible-playbook -i provisioner/hosts provisioner/site.yml --syntax-check
-fab test:"--cov"
+make test ARGS="--cov"
 
 # Cleanup
 test ! $CI && dropdb hello_world
